update: 移除部分无用代码,拆分split功能函数
